Removed embedded form and made custom
Updated the html and the js to remove the embeddded form and replace it with a custom one. The custom form is set up to send via api to whereever we will want. Currently sends to the Make.com webhook and runs through the automation and gets saved to a google drive sheet